+++ src/main/java/com/thunder11/scuad/jobposting/domain/type/AnalysisType.java
@@ -4,6 +4,5 @@ public enum AnalysisType {
     EVALUATION,
     RESUME,
     PORTFOLIO,
-    COMPARISON,
-    ALL
+    COMPARISON
 }
